crypto: Move TapTweakHashExt trait to crypto crate
The TapTweakHashExt trait functionality is only used in crypto::sighash
and bitcoin::taproot. Since this functionality can't be located in
taproot-primitives due to circular dependencies, and it can't be in
bitcoin since that blocks crypto use, it's best located in crypto
itself.

Add taproot-primitives dep to crypto.
Move TapTweakHashExt trait to crypto.
Re-export TapTweakHashExt from bitcoin::taproot.
