+++ crypto/src/sighash.rs
@@ -13,7 +13,12 @@ use core::{fmt, str};
 
 #[cfg(feature = "arbitrary")]
 use arbitrary::{Arbitrary, Unstructured};
+use hashes::{sha256t, HashEngine as _};
+use taproot_primitives::{TapNodeHash, TapTweakHash, TapTweakTag};
 
+use crate::key::UntweakedPublicKey;
+
+#[rustfmt::skip]                // Keep public re-exports separate.
 #[doc(no_inline)]
 pub use self::error::{
     InvalidSighashTypeError, NonStandardSighashTypeError, SighashTypeParseError,
@@ -223,6 +228,35 @@ impl From<EcdsaSighashType> for TapSighashType {
             EcdsaSighashType::NonePlusAnyoneCanPay => Self::NonePlusAnyoneCanPay,
             EcdsaSighashType::SinglePlusAnyoneCanPay => Self::SinglePlusAnyoneCanPay,
         }
+    }
+}
+
+/// Extension functionality for the [`TapTweakHash`] type.
+pub trait TapTweakHashExt {
+    /// Constructs a new BIP-0341 [`TapTweakHash`] from key and Merkle root. Produces `H_taptweak(P||R)` where
+    /// `P` is the internal key and `R` is the Merkle root.
+    fn from_key_and_merkle_root<K: Into<UntweakedPublicKey>>(
+        internal_key: K,
+        merkle_root: Option<TapNodeHash>,
+    ) -> Self;
+}
+
+impl TapTweakHashExt for TapTweakHash {
+    fn from_key_and_merkle_root<K: Into<UntweakedPublicKey>>(
+        internal_key: K,
+        merkle_root: Option<TapNodeHash>,
+    ) -> Self {
+        let internal_key = internal_key.into();
+        let mut eng = sha256t::Hash::<TapTweakTag>::engine();
+        // always hash the key
+        eng.input(&internal_key.serialize().0);
+        if let Some(h) = merkle_root {
+            eng.input(h.as_ref());
+        } else {
+            // nothing to hash
+        }
+        let inner = sha256t::Hash::<TapTweakTag>::from_engine(eng);
+        Self::from_byte_array(inner.to_byte_array())
     }
 }
